Creating brush edit redux action
